requirements.txt
# Core Django packages
Django>=3.2,<4.0
djangorestframework>=3.14.0
django-taggit>=4.0.0
python-dotenv>=0.21.0
Pillow>=9.0.0
django-cors-headers>=4.0.0

# Production database (uncomment when deploying)
# mysqlclient>=2.1.0